+++ crates/messaging/affinidi-messaging-didcomm-v1/src/protocols/mod.rs
@@ -6,4 +6,6 @@
 //! top-level members.
 
 pub mod basic_message;
+pub mod coordinate_mediation;
 pub mod forward;
+pub mod message_pickup;